requirements.txt
flask_cors
werkzeug
wsgiserver
psycopg2-binary